Add base16-solarized theme
This theme is a 256 color variant and it uses the color
codes of the base16-solarized (256 color) terminal/shell theme:
https://github.com/chriskempson/base16
https://github.com/chriskempson/base16-shell

It is based on the original tmux-colors-solarized light theme with some
color codes changed (to map solarized base2, base00, orange, base1).

In fact, this theme may be used in combination with any base16 256 color
terminal/shell theme. But it will probably look a bit 'burnt' (i.e.
solarized).  It better matches the base16-solarized-light and
base16-solarized-dark terminal/shell color themes.